[feat] 1. test_ssd_read 및 test_ssd_write 파일 통합
